refactor(dag-nodes): SEC-007 split the tool node by responsibility, not by baselining it
Containing the node pushed `index.ts` to 313 lines, over the 300-line ceiling.
The anti-monolith rule says split by responsibility rather than record a new
baseline, and there are two genuine responsibilities to lift out:

- `containment.ts` — deciding the security boundary. A boundary buried in an
  executor is a boundary nobody reviews.
- `tool-factories.ts` — WHICH builtins exist and HOW each binds to the root, a
  registry concern separate from executing a node.

`index.ts` is now 219 lines. No behaviour change: 26 passed, typecheck clean,
file-size scan green.
